The astral-go pin moves to the astral field types
astral-go moved four wire fields to astral types and registered
auth.Action as mod.auth.action and the op parameter as
routing.op_param. The pin follows, and the call sites convert at the
boundary: Query.QueryString is String32, services.Update.Available and
nat.ConsumeHoleSignal.Ok are Bool, log.Entry.Level is Uint8, and
routing.OpSpec.Name is String32.

mod.auth.sudo_action and mod.shell.shell_action derive a Blueprint with
no change here: the base they embed is a registered type now.

Measured on a node loading every module: of 175 registered types, one
holds a non-astral field — object_id.sha256, a primitive with a
hand-written codec — where 34 did before this series.

+++ mod/log/src/module.go
@@ -32,7 +32,7 @@ func (mod *Module) LogEntryFilter(entry *log.Entry) bool {
 	if lvl == nil {
 		return entry.Level <= DefaultLogLevel
 	}
-	return entry.Level <= *lvl
+	return uint8(entry.Level) <= *lvl
 }
 
 func (mod *Module) Run(ctx *astral.Context) error {
